add Server.js, add self-invoking function on index.js, reformat folders and path to files
