--- src/data/apiData.json
@@ -1,20 +1,0 @@
-[
-  {
-    "id": "1",
-    "username": "Vadzim",
-    "age": "35",
-    "hobbies": "Exercitation fugiat ad sint nostrud in et laboris. Cupidatat non incididunt ea commodo elit minim ex ea duis exercitation irure do. Ex occaecat sint ea nostrud ea ea ut excepteur ea nulla ullamco amet culpa. Non ad occaecat labore sint cupidatat."
-  },
-  {
-    "id": "2",
-    "username": "Jakub",
-    "age": "34",
-    "hobbies": "Veniam exercitation do enim veniam laboris in fugiat ullamco consequat laboris cillum velit. Cillum sit eiusmod nisi adipisicing deserunt adipisicing proident irure. Nulla nisi esse exercitation aliqua officia deserunt pariatur nulla."
-  },
-  {
-    "id": "3",
-    "username": "Jan",
-    "age": "33",
-    "hobbies": "Nostrud aute commodo esse officia laboris anim eiusmod Lorem ut aute est. Enim commodo esse qui consectetur sit duis anim. Dolor ipsum cillum sit ad ad qui culpa. Enim id mollit nisi esse cillum cillum amet id dolore do nulla ullamco. Irure consequat officia voluptate amet mollit cupidatat aliquip anim proident nostrud ipsum aliqua id. Et in cillum nisi et ullamco pariatur officia eu irure consectetur amet excepteur eiusmod. Labore incididunt enim velit ipsum. "
-  }
-]
